(feat) Refactor Nav - dynamic side-links & search
- Split Nav into 3 (Nav, NavTop, NavSide)
- NavTop only has logout and home + search bar
- NavSide will now change the list of links on the side
based on what component is currently on the page
